backend/requirements/base.txt
bcrypt==4.0.1
pydantic[email]==2.3.0
pydantic-settings==2.0.3
fastapi==0.103.2
uvicorn[standard]==0.22.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.0.2
pdfminer.six==20221105
mangum==0.17.0
python-multipart==0.0.6
